add: ls and make some refactor
