+++ src/helpers/constants.js
@@ -1,4 +1,1 @@
 export const INVALID_INPUT = 'Invalid input';
-export const operationsMap = {
-    up: () => {},
-}
